% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5.R
\name{top5}
\alias{top5}
\title{list ncol x ncol sub matrix}
\usage{
top5(mtx, ncol = 5)
}
\arguments{
\item{mtx}{matrix}

\item{ncol}{submatrix size,default 5}
}
\description{
list ncol x ncol sub matrix
}
\examples{
x<-matrix(rnorm(100),ncol=10)
top5(x)
}

